a rejig to improve naming for package
