+++ data_linter/__init__.py
@@ -1,21 +1,16 @@
 import os
-
-import logging
-from datetime import datetime
 import atexit
 import io
-from goodtables import validate
+from data_linter.logging_functions import upload_log, logging_setup
 
-from logging_functions import upload_log, logging_setup
-
-from functions import (
+from data_linter.validation import (
     load_and_validate_config,
     validate_data,
     get_validator_name,
 )
 
 
-def main():
+def run_validation():
     # set up logging
     log, log_stringio = logging_setup()
 
@@ -26,7 +21,3 @@ def main():
     validate_data(config)
 
     atexit.register(upload_log, body=log_stringio.getvalue(), s3_path=log_path)
-
-
-if __name__ == "__main__":
-    main()
